+++ src/Sql/Sql/Elastic Pools/Model/AzureSqlElasticPoolModel.cs
@@ -140,5 +140,11 @@ namespace Microsoft.Azure.Commands.Sql.ElasticPool.Model
         /// Gets or sets the number of high availability replicas associated with the elastic pool
         /// </summary>
         public int? HighAvailabilityReplicaCount { get; set; }
+
+        /// <summary>
+        /// Gets or sets type of enclave requested on the elastic pool.
+        /// Possible values include: 'Default', 'VBS'
+        /// </summary>
+        public string PreferredEnclaveType { get; set; }
     }
 }