+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Added support for Continuous 7 Days backup mode.
 
 ## Version 1.10.1
 * Updated Azure.Core to 1.31.0.